requirements.txt
torch>=2.0,<3.0
torchvision>=0.15
gradio>=3.34,<4.0
huggingface_hub>=0.19
pillow>=9.0
numpy>=1.23